requirements.txt
python-dotenv
environs~=8.0.0
coverage==5.4
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
Unidecode==1.3.4
uvicorn==0.14.0
psycopg2
requests
python-multipart==0.0.5
Django==4.1
django-environ==0.4.5
django-cors-headers==3.13.0
django-db-connection-pool[postgresql]
django-redis==5.2.0
djangorestframework==3.14.0
PyJWT==2.7.0
djangorestframework-simplejwt
django-filter==23.2
drf-yasg==1.21.5
drf-spectacular==0.26.4

